Fixes the deadlock issue.
Also adds a unit test which tickles the deadlock bug that I think was
causing the issue.  Exceptions + locks = complicated pain.

Also updates all the submodules and adds a line to the TestDockerfile to
ensure that the package list is always updated.  All of these
accompanying changes were necessary to get unit tests to pass again.